feat: semi-stateless authentication
Added a new authentication feature which relies on a short-lived access token that lives on client-side, and a long-lived access token that lives on server-side. This enables the application to regulate and audit user activity. The only cons of this approach is the "enforcement lag", which relies on how long the access token expires before requesting to refresh the access token.

ToDo:
- Add pagination and filters to the "fetch all" resources
- Add "value" options in request validation engine, which whitelists/blacklists certain values
- Add image upload feature